+++ src/Pages/CardListPage/CardListPage.tsx
@@ -12,13 +12,9 @@ export class CardListPage extends React.Component {
     public render(): React.ReactNode {
         return (
             <div className={styles.container}>
-                {ALL_CARD_IDS.map((id, i) => {
-                    const diff = (i === 0 || ALL_CARD_IDS[i].substring(0,3) !== ALL_CARD_IDS[i-1].substring(0,3)) ? 1 : parseInt(ALL_CARD_IDS[i].substring(4))-parseInt(ALL_CARD_IDS[i-1].substring(4));
-                    return (<>
-                        {(diff !== 1 && diff !== 0) && <div>{`${diff-1} missing card(s)`}</div>}
-                        <VDecklistCard cardId={id} options={{type: "showType"}} key={id} />
-                    </>);
-                })}
+                {ALL_CARD_IDS.map((id) => 
+                    <VDecklistCard id={id} options={{type: "showType"}} key={id} />
+                )}
             </div>
         );
     }