+++ src/Pages/PlayerResultsPage/PlayerResultsPage.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 import styles from './PlayerResultsPage.module.css';
 
-import { DEV_STRING_PRE, GET_TOURNAMENT_DECKLIST, GET_TOURNAMENT_ABBR, GET_TOURNAMENT_ID } from '../../Data';
-import { ALL_PLAYERS, GET_PLAYER_RESULTS } from '../../Data/PlayerResults';
-import { Link } from 'react-router-dom';
+import { DEV_STRING_PRE, GET_TOURNAMENT_DECKLIST, GET_TOURNAMENT_ABBR, GET_TOURNAMENT_ID, ALL_PLAYERS, GET_PLAYER_RESULTS } from '../../Data';
 
 export class PlayerResultsPage extends React.Component {
     public override componentDidMount(): void {
@@ -14,15 +13,16 @@ export class PlayerResultsPage extends React.Component {
         return (
             <div className={styles.container}>
                 {ALL_PLAYERS.map((username) => 
-                    <div className={styles.tournamentContainer}>
+                    <div className={styles.tournamentContainer} key={username}>
                         <div className={styles.tournamentHeader}>{username}</div>
                         <div>
                             {GET_PLAYER_RESULTS(username).map((placing) => {
                                 const tournId = GET_TOURNAMENT_ID(GET_TOURNAMENT_ABBR(placing.tournament), placing.date);
                                 const decklist = GET_TOURNAMENT_DECKLIST(tournId, username);
+                                const content = `[${placing.placing}] ${placing.date} ${placing.tournament}`;
                                 return (
-                                    <div className={styles.row}>
-                                        <span>{`[${placing.placing}] ${placing.date} ${placing.tournament}`}</span>
+                                    <div className={styles.row} key={content}>
+                                        <span>{content}</span>
                                         {(decklist !== undefined && decklist.mainDeck.length > 0) &&
                                             <span>
                                                 {" ("}