+++ src/Views/LegendImage/LegendImage.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import styles from './LegendImage.module.css';
 
-import { GET_CARD, GET_CARD_ART } from '../../Data';
+import { GET_CARD, GET_CARD_ART, UNKNOWN_LEGEND_ID } from '../../Data';
 
-interface IProps { printId: string; size: number; imgTitle?: string; subscript?: string; subscriptFontSize?: number; extraStyles?: React.CSSProperties }
+interface IProps { id: string; size: number; imgTitle?: string; subscript?: string; subscriptFontSize?: number; extraStyles?: React.CSSProperties }
 
 export class LegendImage extends React.Component<IProps> {
     public render(): React.ReactNode {
-        const card = GET_CARD(this.props.printId);
+        const card = GET_CARD(this.props.id);
         const imgTitle = this.props.imgTitle ?? card.name.substring(0, card.name.search(","));
         return (
             <div className={styles.imageContainer} style={{width: this.props.size, height: this.props.size, ...(this.props.extraStyles ?? {})}}>
-                {(this.props.printId.length > 0)
-                    ? <img src={GET_CARD_ART(this.props.printId)} className={styles[`position-${this.props.printId}`]} title={imgTitle} alt={imgTitle} />
+                {(this.props.id !== UNKNOWN_LEGEND_ID)
+                    ? <img src={GET_CARD_ART(this.props.id)} className={styles[`position-${this.props.id}`]} title={imgTitle} alt={imgTitle} />
                     : <div className={styles.unknown}>?</div>
                 }
                 <span className={styles.subscript} style={(this.props.subscriptFontSize) ? {fontSize: `${this.props.subscriptFontSize}px`} : {}}>{this.props.subscript}</span>