+++ src/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
-import { AppMainRouter } from './AppMainRouter';
 import reportWebVitals from './reportWebVitals';
+import './index.css';
+
+import { AppMainRouter } from './AppMainRouter';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement